Editor completion: move references to EditorOperations from the new completion
